+++ src/main/java/main/asw/repository/dao/UserDaoImpl.java
@@ -1,4 +1,4 @@
-package main.asw.repository.agentDao;
+package main.asw.repository.dao;
 
 import static com.mongodb.client.model.Filters.eq;
